créer une nouvelle application 'lettings', contenant les modèles 'Address' et 'Letting'.
