+++ oc_lettings_site/settings.py
@@ -28,6 +28,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
+    'lettings',
 ]
 
 MIDDLEWARE = [
