+++ oc_lettings_site/urls.py
@@ -3,10 +3,12 @@ from django.urls import path
 
 from . import views
 
+import lettings.views
+
 urlpatterns = [
     path('', views.index, name='index'),
-    path('lettings/', views.lettings_index, name='lettings_index'),
-    path('lettings/<int:letting_id>/', views.letting, name='letting'),
+    path('lettings/', lettings.views.lettings_index, name='lettings_index'),
+    path('lettings/<int:letting_id>/', lettings.views.letting, name='letting'),
     path('profiles/', views.profiles_index, name='profiles_index'),
     path('profiles/<str:username>/', views.profile, name='profile'),
     path('admin/', admin.site.urls),
